Fix minor clippy warnings (#38)
* Loop over test vectors directly

Address a clippy warning by unpacking over the container elements
directly instead of using a range and index variable. This is more
idiomatic.

* Simplify test assertion

Use the more explicit `is_err` instead of negating `is_ok`.
Addresses a clippy lint.

* Remove redundant struct label in test

Address a clippy warning. When the struct initialization value has
the same name as the member, it's idiomatic to omit the member label.

* Remove unnecessary base64 import

The base64 package is referred to by its full path, so the explict
`use` statement is redundant in contemporary rust compilers.

Addresses a clippy lint.