Remove alloc feature
The `alloc` crate is always available in rust 1.36 and later,
so we don't need a special feature flag to enable it. Instead,
enable the `curve25519-dalek/alloc` feature unconditionally
and rely on building with `--no-default-features` to disable
standard library dependence.

Restores building and running tests in a no_std environment.